+++ pkgs/pool/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fix `Pool.forEach` to ensure all workers complete before the stream is closed, even on error or cancellation.
 * Added an example.
+* Fix a resource leak when `PoolResource.allowRelease` callback throws.
 
 ## 1.5.2